refactor request and response types to apiclients folder
