+++ frontend/src/components/auth/TeacherSignup/TeacherSignupTwo.tsx
@@ -2,7 +2,7 @@ import { useQuery } from "@apollo/client";
 import { Button, VStack, Text, FormControl, FormLabel } from "@chakra-ui/react";
 import React from "react";
 import GET_SCHOOLS from "../../../APIClients/queries/SchoolQueries";
-import { SchoolResponse } from "../../../types/SchoolTypes";
+import { SchoolResponse } from "../../../APIClients/types/SchoolClientTypes";
 import SelectFormInput from "./SelectFormInput";
 import { TeacherSignupProps } from "./types";
 import ErrorMessage from "./ErrorMessage";
