+++ frontend/src/components/common/AddAdminModal.tsx
@@ -27,7 +27,7 @@ import AdminConfirmationMessage from "./Admin/AdminConfirmationMessage";
 import ModalFooterButtons from "./ModalFooterButtons";
 import { PlusOutlineIcon } from "./icons";
 import { ADMIN_PAGE } from "../../constants/Routes";
-import { UserRequest } from "../../types/UserTypes";
+import { UserRequest } from "../../APIClients/types/UserClientTypes";
 import { ADD_USER } from "../../APIClients/mutations/UserMutations";
 import GET_USERS_BY_ROLE from "../../APIClients/queries/UserQueries";
